PLEASE TEST THIS COMMIT!
Rearrange things so ->config methods return a ConfigSet, which is a new container
class for SystemFileSets and CommandSets.

This means that any of tun/bgp/fwall/enc can return any combination of files
or commands, and the upper layers will be cool with it.

There's a lot of changes in this, and as of this commit diff hasn't been tested
at all, although -c works fine, so it should be ok.


git-svn-id: file:///Volumes/Backup Macintosh HD/Users/chris/from_munkyii/ss-repo/svn/funknet-tools/trunk@586 94779f63-8a0d-0410-8a4e-8732221620d0