PF rules. Currently -c mode only.
This uses pf's 'anchor' to do achieve a similar result to the FUNKNET
iptable - provide a distinct place to add FN rules.

The emitted rules need to be written to a file and then fed into pf with
something like

  $ pfctl -a FUNKNET /some/file

As such, the main pf.conf needs to contain

  anchor /some/file

at a convenient point.

( It would be nicer to do pipe the rules straight into pfctl, but each
  successive call to pfctl resets the entire anchor. bah. )

(( I don't fully grok the object layout just yet, may contain bork. ))


git-svn-id: file:///Volumes/Backup Macintosh HD/Users/chris/from_munkyii/ss-repo/svn/funknet-tools/trunk@768 94779f63-8a0d-0410-8a4e-8732221620d0